timeline socket components

+++ wasm/source/EventQueue.cpp
@@ -48,7 +48,9 @@ std::map<double,TEvent*> EventQueue::allNext(glm::vec3 vantage, double time, dou
             if(time_to_run <= time){
                 //printf("queue time: %f\n", time_to_run);
                 if(all_next.find(time_to_run) != all_next.end()){
-                    printf("Events had identical run_time! execution order not guranteed!\n");
+                    //printf("Events had identical run_time! execution order not guranteed!\n");
+                    //Variant(e->serialize()).printFormatted();
+                    //Variant(all_next[time_to_run]->serialize()).printFormatted();
                     all_next = map<double, TEvent*>();
                     all_next[0] = next(vantage, time, info_speed); // fall back to single next to try to recover
                     return all_next ;
@@ -78,26 +80,36 @@ TEvent* EventQueue::addEvent(std::unique_ptr<TEvent> event){
 }
 
 void EventQueue::removeDependencies(TEvent* event){
-    
     // Delete any data following from this write
     if(event->wrote_anchor){
         timeline->objects[event->anchor_id].deleteAfter(event->time);
         event->wrote_anchor = false;
     }
-
     // Delete any events this event spawned
     for(TEvent* s : event->spawned_events){
-        if(!s->deleted){
+        if(s != nullptr && !s->deleted){// TODO s-> deleted can segfault?
             deleteEvent(s);
         }
     }
     event -> spawned_events.clear();
+    // Remove any links on data objects
+    for(int k=0;k<event->read.size();k++){
+        if(event->read[k] != nullptr){
+            for(int j=0;j<event->read[k]->readers.size();j++){
+                if(event->read[k]->readers[j].first == event){
+                    event->read[k]->readers[j].first = nullptr ;
+                }
+            }
+        }
+    }
+    event->read.clear();
+    
+    timeline->collisions.removeRequests(event);
 }
 
 void EventQueue::deleteEvent(TEvent* event){
     //printf("Deleting event at time %f \n", event->time);
     removeDependencies(event);
-    timeline->collisions.onDelete(event);
     event->deleted = true;
     event->spawner = nullptr ;
 }
@@ -110,21 +122,46 @@ void EventQueue::rerunEvent(TEvent* event){
 
 // Clears out all events before the given time
 void EventQueue::clearHistoryBefore(double clear_time){
-    // mark deleted and clear all pointers to events about to be newly removed
+    // mark deleted
     for(int k=0;k<events.size();k++){
         if(events[k].get() != nullptr && !events[k]->deleted && events[k]->time < clear_time && !events[k]->run_pending ){
-            timeline->collisions.onDelete(events[k].get()); // remove pointers in collision system
+            timeline->collisions.removeRequests(events[k].get()); // remove pointers in collision system
             events[k]->deleted = true;
-            events[k]->spawner = nullptr ;
-            // Remove the link to this as a spawner of future events that may not be deleted
-            for(TEvent* s : events[k]->spawned_events){
-                s->spawner = nullptr ;
-            }
         }
     }
-    //wipe the data for all deleted events
+    //wipe the data for all deleted events and clear all pointers to events about to be newly removed
     for(int k=0;k<events.size();k++){
         if(events[k].get() != nullptr && events[k]->deleted){
+
+            if(events[k]->spawner != nullptr){
+                // clear reference in parent's spawned event list to this
+                for(int s=0;s< events[k]->spawner->spawned_events.size();s++){
+                    if(events[k]->spawner->spawned_events[s] == events[k].get()){
+                        events[k]->spawner->spawned_events[s] = nullptr;
+                    }
+                }
+                // clear link to spawner
+                events[k]->spawner = nullptr ;
+            }
+            // Remove the link to this as a spawner of future events that may not be deleted
+            for(TEvent* s : events[k]->spawned_events){
+                if(s != nullptr){
+                    s->spawner = nullptr ;
+                }
+            }
+
+            // Remove links to event on objects that it read
+            for(int i=0;i<events[k]->read.size();i++){
+                if(events[k]->read[i] != nullptr){
+                    for(int j=0;j<events[k]->read[i]->readers.size();j++){
+                        if(events[k]->read[i]->readers[j].first == events[k].get()){
+                            events[k]->read[i]->readers[j].first = nullptr ;
+                        }
+                    }
+                }
+            }
+            events[k]->read.clear();
+
             events[k].reset(); // actually free the data in the event
         }
     }
@@ -136,10 +173,38 @@ std::vector<TEvent*> EventQueue::getBase(double time){
     for(int k=0;k<events.size();k++){
         if(events[k].get() != nullptr && !events[k]->deleted && events[k]->time > time){
             TEvent* spawner = events[k]->spawner ;
-            if(spawner == nullptr || spawner->time < time){
+            if(spawner == nullptr || spawner->time <= time || spawner->deleted){
                 base.push_back(events[k].get());
             }
         }
     }
+    /*
+    if(base.size() !=0 && base.size() != 10){
+        printf("Events at get base current time: %f last_clear_time %f  base time: %f\n", timeline->current_time, timeline->last_clear_time, time); 
+        for(int k=0;k<events.size();k++){
+             if(events[k].get() != nullptr){
+                printf(" %d hash : %d, deleted : %s has spawner: %s\n", k, 
+                Variant(events[k]->serialize()).hash(), 
+                events[k]->deleted ? "true" : "false",
+                events[k]->spawner == nullptr ? "false" : "true");
+                if(events[k]->spawner != nullptr){
+                    printf("spawner hash: %d\n", Variant(events[k]->spawner->serialize()).hash());
+                }
+                Variant(events[k]->serialize()).printFormatted();
+
+                bool base= false;
+                if(events[k].get() != nullptr && !events[k]->deleted && events[k]->time > time){
+                    TEvent* spawner = events[k]->spawner ;
+                    if(spawner == nullptr || spawner->time <= time || spawner->deleted){
+                        base = true ;
+                    }
+                }
+
+                printf("Final verdict for Base event: %s\n", base ? "true" : "false");
+             }
+        }
+        printf("Generated Base events size: %d\n", (int)base.size());
+    }*/
+
     return base ;
 }
